refactor: extract grammarCompletion.ts from grammarMatcher.ts
Move the completion system (matchGrammarCompletion, GrammarCompletionResult,
and all supporting completion functions) into a dedicated grammarCompletion.ts
module. Update exports in index.ts, fix imports in nfaCompletion.ts and
testUtils.ts, and update documentation references across README.md,
NFA_README.md, and actionGrammar.md.
